# public/swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Next.js Documentation App",
    "version": "1.0.0",
    "description": "API for user authentication, progress tracking, and profile management"
  },
  "paths": {
    "/api/users/login": {
      "post": {
        "summary": "User login",
        "tags": ["Users"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "example": "user@example.com" },
                  "password": { "type": "string", "example": "mypassword" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Logged in successfully" },
          "400": { "description": "Invalid credentials or user does not exist" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/api/users/logout": {
      "get": {
        "summary": "User logout",
        "tags": ["Users"],
        "responses": {
          "200": { "description": "Logged out successfully" }
        }
      }
    },
    "/api/users/register": {
      "post": {
        "summary": "Register a new user",
        "tags": ["Users"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string", "example": "JohnDoe" },
                  "email": { "type": "string", "example": "user@example.com" },
                  "password": { "type": "string", "example": "mypassword" }
                },
                "required": ["username", "email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "User registered successfully" },
          "400": { "description": "Validation error or user already exists" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/api/users/progress": {
      "get": {
        "summary": "Get user progress",
        "tags": ["Progress"],
        "responses": {
          "200": { "description": "List of progress pages" },
          "500": { "description": "Server error" }
        }
      },
      "post": {
        "summary": "Mark a page as read",
        "tags": ["Progress"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "page": { "type": "string", "example": "page3" }
                },
                "required": ["page"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Page marked as read" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/api/users/me": {
      "get": {
        "summary": "Get current user profile",
        "tags": ["Profile"],
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "User profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "_id": { "type": "string" },
                        "username": { "type": "string" },
                        "email": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "404": { "description": "User not found" },
          "500": { "description": "Server error" }
        }
      },
      "put": {
        "summary": "Update current user profile",
        "tags": ["Profile"],
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string", "example": "JohnDoe" },
                  "email": { "type": "string", "example": "user@example.com" }
                },
                "required": ["username", "email"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "_id": { "type": "string" },
                        "username": { "type": "string" },
                        "email": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Username and email required" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "User not found" },
          "500": { "description": "Server error" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token"
      }
    }
  }
}